Create README.md
When a user clicks an already booked slot (yellow), a message saying that the slot has already been taken will appear. When a user clicks on an available slot (green), he should be provided with an option to enter his email address and click book or cancel. When "book" is clicked, the corresponding reservation in the database is updated with booking flag = 1 and the user email is entered. The slot now becomes unavailable (yellow).